add sign in info to homepage
